mcp-server/requirements.txt
azure-functions==1.25.0b3
